Add options page for setting custom elements
