+++ source/android/mobile/samples/v1.0/CustomElement.json
@@ -1,0 +1,11 @@
+{
+	"$schema": "http://adaptivecards.io/schemas/adaptive-card.json",
+	"type": "AdaptiveCard",
+	"version": "1.0",
+	"body": [
+		{
+            "type": "blah",
+            "secret": "This is a custom element"
+		}
+	]
+}
