+++ source/android/mobile/src/main/java/io/adaptivecards/adaptivecardssample/CustomObjects/Media/OnlineImageLoader.java
@@ -1,0 +1,35 @@
+package io.adaptivecards.adaptivecardssample.CustomObjects.Media;
+
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+
+import java.io.IOException;
+import java.net.URISyntaxException;
+
+import io.adaptivecards.renderer.GenericImageLoaderAsync;
+import io.adaptivecards.renderer.IOnlineImageLoader;
+import io.adaptivecards.renderer.http.HttpRequestHelper;
+import io.adaptivecards.renderer.http.HttpRequestResult;
+
+public class OnlineImageLoader implements IOnlineImageLoader
+{
+    @Override
+    public HttpRequestResult<Bitmap> loadOnlineImage(String url, GenericImageLoaderAsync loader) throws IOException, URISyntaxException
+    {
+        String catImnageUri = "http://adaptivecards.io/content/cats/1.png";
+        byte[] bytes = HttpRequestHelper.get(catImnageUri);
+        if (bytes == null)
+        {
+            throw new IOException("Failed to retrieve content from " + catImnageUri);
+        }
+
+        Bitmap bitmap = BitmapFactory.decodeByteArray(bytes, 0, bytes.length);
+
+        if (bitmap == null)
+        {
+            throw new IOException("Failed to convert content to bitmap: " + new String(bytes));
+        }
+
+        return new HttpRequestResult<>(bitmap);
+    }
+}
