+++ source/android/mobile/src/main/java/io/adaptivecards/adaptivecardssample/CustomObjects/Media/SvgImageLoader.java
@@ -1,0 +1,64 @@
+package io.adaptivecards.adaptivecardssample.CustomObjects.Media;
+
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.graphics.drawable.Drawable;
+
+import com.pixplicity.sharp.Sharp;
+
+import java.io.IOException;
+import java.net.URISyntaxException;
+import java.net.URLDecoder;
+
+import io.adaptivecards.adaptivecardssample.ImageUtil;
+import io.adaptivecards.objectmodel.AdaptiveBase64Util;
+import io.adaptivecards.objectmodel.CharVector;
+import io.adaptivecards.renderer.GenericImageLoaderAsync;
+import io.adaptivecards.renderer.IOnlineMediaLoader;
+import io.adaptivecards.renderer.IResourceResolver;
+import io.adaptivecards.renderer.Util;
+import io.adaptivecards.renderer.http.HttpRequestResult;
+
+public class SvgImageLoader implements IResourceResolver
+{
+    @Override
+    public HttpRequestResult<Bitmap> resolveImageResource(String uri, GenericImageLoaderAsync genericImageLoaderAsync) throws IOException, URISyntaxException
+    {
+        Bitmap bitmap;
+        String dataUri = AdaptiveBase64Util.ExtractDataFromUri(uri);
+        CharVector decodedDataUri = AdaptiveBase64Util.Decode(dataUri);
+        byte[] decodedByteArray = Util.getBytes(decodedDataUri);
+        bitmap = BitmapFactory.decodeByteArray(decodedByteArray, 0, decodedByteArray.length);
+
+        return new HttpRequestResult<>(bitmap);
+    }
+
+    @Override
+    public HttpRequestResult<Bitmap> resolveImageResource(String uri, GenericImageLoaderAsync genericImageLoaderAsync, int maxWidth) throws IOException, URISyntaxException
+    {
+        Bitmap bitmap;
+        String dataUri = AdaptiveBase64Util.ExtractDataFromUri(uri);
+        CharVector decodedDataUri = AdaptiveBase64Util.Decode(dataUri);
+
+        if (uri.startsWith("data:image/svg")) {
+            String svgString = AdaptiveBase64Util.ExtractDataFromUri(uri);
+            String decodedSvgString = URLDecoder.decode(svgString, "UTF-8");
+            Sharp sharp = Sharp.loadString(decodedSvgString);
+            Drawable drawable = sharp.getDrawable();
+            bitmap = ImageUtil.drawableToBitmap(drawable, maxWidth);
+        }
+        else
+        {
+            try
+            {
+                return genericImageLoaderAsync.loadDataUriImage(uri);
+            }
+            catch (Exception e)
+            {
+                return new HttpRequestResult<>(e);
+            }
+        }
+
+        return new HttpRequestResult<>(bitmap);
+    }
+}
