+++ source/android/mobile/src/main/res/layout/content_main_activity_adaptive_cards_sample.xml
@@ -180,6 +180,123 @@
                                 android:inputType="textMultiLine" />
                         </LinearLayout>
                     </ScrollView>
+
+                </LinearLayout>
+
+                <LinearLayout
+                    android:id="@+id/options"
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:orientation="vertical">
+
+                    <ScrollView
+                        android:layout_width="match_parent"
+                        android:layout_height="match_parent">
+
+                        <LinearLayout
+                            android:layout_width="match_parent"
+                            android:layout_height="wrap_content"
+                            android:orientation="vertical" >
+
+                            <Switch
+                                android:id="@+id/customActions"
+                                android:layout_width="match_parent"
+                                android:layout_height="wrap_content"
+                                android:text="Use custom actions" />
+
+                            <TextView
+                                android:id="@+id/cardsCustomActions"
+                                android:layout_width="match_parent"
+                                android:layout_height="wrap_content"
+                                android:paddingLeft="15dp"
+                                android:text="CustomActionRenderingTest.json"
+                                android:visibility="gone" />
+
+                            <Switch
+                                android:id="@+id/customElements"
+                                android:layout_width="match_parent"
+                                android:layout_height="wrap_content"
+                                android:text="Use custom elements" />
+
+                            <TextView
+                                android:id="@+id/cardsCustomElements"
+                                android:layout_width="match_parent"
+                                android:layout_height="wrap_content"
+                                android:paddingLeft="15dp"
+                                android:text="CustomElement.json"
+                                android:visibility="gone" />
+
+                            <Switch
+                                android:id="@+id/customFeatureReg"
+                                android:layout_width="match_parent"
+                                android:layout_height="wrap_content"
+                                android:text="Register Feature Registration" />
+
+                            <TextView
+                                android:id="@+id/cardsCustomFeature"
+                                android:layout_width="match_parent"
+                                android:layout_height="wrap_content"
+                                android:paddingLeft="15dp"
+                                android:text="Element.Requires.json"
+                                android:visibility="gone" />
+
+                            <Switch
+                                android:id="@+id/svgSupport"
+                                android:layout_width="match_parent"
+                                android:layout_height="wrap_content"
+                                android:text="Support SVG images" />
+
+                            <TextView
+                                android:id="@+id/cardsSvgSupport"
+                                android:layout_width="match_parent"
+                                android:layout_height="wrap_content"
+                                android:paddingLeft="15dp"
+                                android:text="Image.SvgDataUri.json"
+                                android:visibility="gone" />
+
+                            <Switch
+                                android:id="@+id/customImageLoader"
+                                android:layout_width="match_parent"
+                                android:layout_height="wrap_content"
+                                android:text="Custom Image Loaders" />
+
+                            <TextView
+                                android:id="@+id/cardsCustomImageLoader"
+                                android:layout_width="match_parent"
+                                android:layout_height="wrap_content"
+                                android:paddingLeft="15dp"
+                                android:text="Image.ResourceResolver.json"
+                                android:visibility="gone" />
+
+                            <Switch
+                                android:id="@+id/customMediaLoader"
+                                android:layout_width="match_parent"
+                                android:layout_height="wrap_content"
+                                android:text="Use custom media loader" />
+
+                            <TextView
+                                android:id="@+id/cardsCustomMediaLoader"
+                                android:layout_width="match_parent"
+                                android:layout_height="wrap_content"
+                                android:paddingLeft="15dp"
+                                android:text="Media.json"
+                                android:visibility="gone" />
+
+                            <Switch
+                                android:id="@+id/onlineImageLoader"
+                                android:layout_width="match_parent"
+                                android:layout_height="wrap_content"
+                                android:text="OnlineImageLoader" />
+
+                            <TextView
+                                android:id="@+id/cardsCustomOnlineImageLoader"
+                                android:layout_width="match_parent"
+                                android:layout_height="wrap_content"
+                                android:paddingLeft="15dp"
+                                android:text="Image.json"
+                                android:visibility="gone" />
+                        </LinearLayout>
+                    </ScrollView>
                 </LinearLayout>
             </FrameLayout>
         </LinearLayout>
